replace Vec3 with Vector3f in most places, except mixin. (optimization of memory and no new instance created every time Vector method used)

+++ visor-core/src/main/java/me/phoenixra/visor/core/client/settings/VRClientSettings.java
@@ -195,8 +195,8 @@ public class VRClientSettings {
     public static void calibrateHeight() {
 
         VRClientSettings.setPlayerHeight(
-                (float) ClientContext.rawPoseHandler.getHmdData()
-                        .getPivotHistory().averagePosition(0.5D).y
+                ClientContext.rawPoseHandler.getHmdData()
+                        .getPivotHistory().averagePosition(0.5f).y
         );
         int i = (int) (Math.round(100.0D
                 * VRClientSettings.getPlayerHeight()
